+++ qualtran/bloqs/block_encoding/lcu_block_encoding.py
@@ -30,8 +30,8 @@ from qualtran import (
     SoquetT,
 )
 from qualtran.bloqs.block_encoding.block_encoding_base import BlockEncoding
-from qualtran.bloqs.block_encoding.lcu_select_and_prepare import SelectOracle
 from qualtran.bloqs.bookkeeping import Partition
+from qualtran.bloqs.multiplexers.select_base import SelectOracle
 from qualtran.bloqs.state_preparation.prepare_base import PrepareOracle
 from qualtran.symbolics import SymbolicFloat
